changed to struc floats to doubles

+++ s_fish_school.c
@@ -17,16 +17,16 @@
 
 int main(int argc, char *argv[])
 {
-	float max_delta;
+	double max_delta;
 	float bary_numer; // technically should have x and y for barycentre
 	// but project requirements do not give that equation
 	float bary_denom;
 	struct fish{
-		float x;
-		float y;
-		float w;
-		float euc_dist;
-		float delta_f;
+		double x;
+		double y;
+		double w;
+		double euc_dist;
+		double delta_f;
 	};
 	srand(time(NULL));
